Move the meta data fields back up in the news/events module (see #1727)
Description
-----------

| Q                | A
| -----------------| ---
| Fixed issues     | Fixes #1666
| Docs PR or issue | -

<!--
Bugfixes should be based on the 4.4 or 4.9 branch and features on the master
branch. Select the correct branch in the "base:" drop-down menu above.

Replace this notice with a short README for your feature/bugfix. This will help
people to understand your PR and can be used as a start for the documentation.
-->

Commits
-------

3f3f240d Move the meta data fields back up in the news/events module