Explicitly register the model classes (see #1951)
Description
-----------

I have realized that using `Model::getRelated()` still relies on the old Contao 3 class autoloader to find the related model class. This is because `Model::getClassFromTable()` only converts `tl_page` to `PageModel` and not to `Contao\PageModel` and therefore we still have "composerized" classes in Contao 4.10 when we really should not anymore. 😄 

<img width="724" alt="" src="https://user-images.githubusercontent.com/1192057/88038330-5c9a8080-cb46-11ea-9976-6441d9ab906a.png">

This PR explicitly registers the fully qualified model class names.

Commits
-------

9857c385 Explicitly register the model classes
3cb436fd Append to TL_MODELS instead of overwriting it
c88a3552 Trigger a deprecation for unregistered tables