Deprecate the insert tag flag "|absolute" (see #3146)
Description
-----------

While discussing https://github.com/contao/contao/pull/3142#issuecomment-872155666 I noticed, that we have an insert tag flag named `absolute`. This is the only flag that is used as an argument AFAIK. All other flags are related to caching or postprocessing the output.

As I understand the insert tag concept this is against the design and IMO very confusing. 
Therefore we should deprecate it IMO.

Is this a feature and still allowed for 4.12 ? Or should I rebase it on 4.9 ?

Commits
-------

005f1a0c Deprecate insert tag flag absolute
6ab8574d Coding style
4dcb618e Update calendar, faq and news insert tags
5033c8bd Merge branch '4.12' into fix/insert-tag-absolute-flag