Add primary image of news and events to meta data (see #5849)
Description
-----------

In #2801 we introduced the possibility for search results to automatically show the `primaryImageOfPage` for the `WebPage` meta data. However, this is currently not utilised (by default) in Contao. This PR would add the teaser image of news and events as the `primaryImageOfPage` and thus the teaser image would automatically be displayed in the search results (after https://github.com/contao/contao/pull/5848 gets merged upstream).

_Note:_ since this is a simple template change, we could also introduce this as a bugfix in Contao 4.13? wdyt?

Commits
-------

f60d7883 add primary image of news and events
084c50e3 Merge remote-tracking branch 'origin/5.x' into add-news-events-primar…