Compare start and stop dates as numbers instead of strings (see #6084)
Description
-----------

With quotes around the start or stop time, MySQL will compare the condition as strings instead of numbers. This means a string `823647600` (ca. 1990) is not before `1685028960` (ca. today), because each letter is compared starting with 8 vs 1.

This is normally not an issue if the start/stop date is within a reasonable range (same length), which is usually the case if you set a start or stop date a few days/weeks/months away from today. I had an issue with members because we used the `tl_member.start` field as "client since", and the "lost password" module was unable to find a member because of start date mismatch.

Commits
-------

6424e29b Fixed comparing start and stop dates as strings instead of numbers