Add `authorModel` to the news templates (see #6145)
Description
-----------

In https://github.com/contao/contao/pull/3119 we introduced an `'authorModel'` entry to `ModuleNews::getMetaFields`. However, this was then actually never passed on to the template and thus `'authorModel'` is completely unused. While the JSON-LD does output an author as this was later on fixed in https://github.com/contao/contao/pull/3161 - but the `'authorModel'` remains unused now. This fixes that by actually passing the `authorModel` to the news template (which can be useful, if you need just the author's name somewhere, without the "by" from the translation).

Commits
-------

ad970f5b add authorModel to template