Handle empty news feed items in the news feed controller (see #8389)
Description
-----------

`TransformArticleForFeedEvent::getItem()` allows to return `null` - i.e. an event listener could decide to not set a feed item - or set it to `null`.

However, the `NewsFeedController` currently does not take this into account (`Feed::add()` does not allow `null`). This PR fixes that.

This PR also simplifies the code style a bit, by eliminating an `if` block.

Commits
-------

16dc56f2 Take empty item into consideration